+++ lib/google.js
