+++ pages/api/youtube.js
@@ -1,6 +1,6 @@
 import { google } from 'googleapis';
 
-import googleAuth from '@/lib/google/auth';
+import googleAuth from '@/lib/google';
 
 export default async function handler(_, res) {
   const auth = await googleAuth.getClient();
